+++ lib/Integrations/YoastDuplicatePost.php
@@ -25,6 +25,12 @@ class YoastDuplicatePost {
 
 		// Skip WPML to Unbabble migration of ubb_source to meta for rewrite republish copies.
 		\add_filter( 'ubb_wpml_migrate_skip_source', [ $this, 'skip_wpml_migrate_source' ], 10, 3 );
+
+		// Hide rewrite republish posts from being linked in other posts.
+		\add_filter( 'ubb_possible_links_filter_sql', [ $this, 'hide_rewrite_republish_from_linking' ], 10, 2 );
+
+		// Remove ubb_source from rewrite republish copies when saved.
+		\add_action( 'wp_insert_post', [ $this, 'clean_republish_copies' ], 10, 2 );
 	}
 
 	public function exclude_ubb_source( array $meta_keys ) : array {
@@ -470,5 +476,46 @@ class YoastDuplicatePost {
 		}
 
 		return $skip;
+	}
+
+	/**
+	 * Hide rewrite republish posts from being linked in other posts.
+	 *
+	 * @since Unreleased
+	 *
+	 * @param string $sql
+	 * @return string
+	 */
+	public function hide_rewrite_republish_from_linking( string $sql ) : string {
+		global $wpdb;
+
+		$sql .= " AND P.ID NOT IN (
+			SELECT post_id
+			FROM {$wpdb->postmeta}
+			WHERE meta_key = '_dp_is_rewrite_republish_copy'
+			AND meta_value = '1'
+		)";
+
+		return $sql;
+	}
+
+	/**
+	 * Remove ubb_source from rewrite republish copies when saved.
+	 *
+	 * @since Unreleased
+	 *
+	 * @param int     $post_id
+	 * @param WP_Post $post
+	 */
+	public function clean_republish_copies( $post_id, $post ) : void {
+		$permissions_helper = new \Yoast\WP\Duplicate_Post\Permissions_Helper();
+		if (
+			! $post instanceof WP_Post
+			|| ! $permissions_helper->is_rewrite_and_republish_copy( $post )
+		) {
+			return;
+		}
+
+		delete_post_meta( $post_id, 'ubb_source' );
 	}
 }